requirements.txt
Django>=5.1.4
djangorestframework>=3.15.2
django-cors-headers>=4.6.0
psycopg2-binary>=2.9.10
python-dotenv>=1.0.1
python-socketio>=5.12.1
channels>=4.2.0
eventlet>=0.38.2